Fix _NET_DESKTOP_VIEWPORT
XChangeProperty seems to expect that the data array has elements of
length long int. I'm yet unsure whether it must be 'long' or 'unsigned
long', however this works in any case.

This should solve the following issue with polybar:
https://github.com/jaagr/polybar/issues/1097

+++ src/ewmh.cpp
@@ -119,7 +119,7 @@ void ewmh_init() {
     ewmh_update_wmname();
 
     /* init atoms that never change */
-    int buf[] = { 0, 0 };
+    long buf[] = { 0, 0 };
     XChangeProperty(g_display, g_root, g_netatom[NetDesktopViewport],
         XA_CARDINAL, 32, PropModeReplace, (unsigned char *) buf, LENGTH(buf));
 }
